Add support for python 3
This plugin currently crashes under python 3 for several reasons related
to string encoding.

In python 3 the unicode type no longer exists and is simply the type
str. As of Ansible 2.5 Ansible now supports the use of python 3 and this
module currently fails if invoked under python 3 because the unicode
function no longer exists.

Six provides a compatibility layer to support both python 2.7 and python
3. Using six.text_type works safely in both python versions.

Additionally the encode function currently returns the raw string if no
encoding is set which results in a string returned to ansible prefixed
with a b which breaks compatibility with dictionary processing in
Ansible. Instead this now defaults to ASCII to return ASCII strings.